trustee_code_of_conduct.md - drop "-ing"s on description text
Drop the -ing on various words (I believe this is called the present
progressive) which sound inconsistent when combined with the phrase
"It is the responsibility of the trustees to".

For example:

  It is the responsibility of the trustees to respect confidentiality
  - understanding what confidentiality means in practice for the
  Hacklab, its board and the individuals involved with it.

Does make sense when read in isolation and in full, but jars a little
when read in the bulleted form, where context takes the initial clause
and the bold title out of the flow:

  It is the responsibility of the trustees to [...] understanding what
  confidentiality means in practice for the Hacklab, its board and the
  individuals involved with it.

And especially so when the description is broken up, resulting in
sentences like:

  Not gaining materially or financially unless specifically authorised
  to do so.

This change is a suggested alternative.